+++ src/ejb/UtilisateurImpl.java
@@ -53,7 +53,7 @@ public class UtilisateurImpl {
 
     }
     
-    public void lierUtilisateur(Long Id, Utilisateur utilisateur){
+    public void lierUtilisateur(int Id, Utilisateur utilisateur){
     	Etudiant e = em.find(Etudiant.class, Id);
     	utilisateur.setEtudiant(e);
     	utilisateur.setLie("Oui");
